+++ CHANGELOG.md
@@ -7,7 +7,9 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 
 ### Fixed
 * [`jsx-indent-props`]: Apply indentation when using brackets ([#2826][] @Moong0122)
+* [`jsx-handler-names`]: Skip inline handlers when checkInlineFunction=false ([#2833][] @onigoetz)
 
+[#2833]: https://github.com/yannickcr/eslint-plugin-react/issues/2833
 [#2826]: https://github.com/yannickcr/eslint-plugin-react/issues/2826
 
 ## [7.21.4] - 2020.10.09